My Jobs Page Frontend
This updates the My Jobs page the recruiter sees more completed. Each job that the recruiter has made is shown here as its own field. You can edit or delete the job posting here. You are able to open the description and see the applications to this job and open a new window to see more information on the applications. This application window for the job shows more details on the person who applied as well as their message. The recruiter can contact them from this page.